+++ src/main/java/com/project/bookuluv/app/admin/controller/ADMController.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
@@ -51,5 +52,12 @@ public class ADMController {
         List<Article> articleList = this.articleService.getAll();
         model.addAttribute("articleList", articleList);
         return "article_list";
+    }
+
+    @GetMapping(value = "/detail/{id}")
+    private String detail(Model model, @PathVariable("id") Integer id) {
+        Article article = this.articleService.getById(id);
+        model.addAttribute("article", article);
+        return "article_detail";
     }
 }
